+++ src/ignitus-Dashboard/ignitus-Opportunity/ignitus-OpportunityDetails/constants.ts
